Fix circular import issue breaking tests
