+++ src/state/session/__tests__/session-test.ts
@@ -12,6 +12,9 @@ jest.mock('jwt-decode', () => ({
 
 jest.mock('../../birthdate')
 jest.mock('../../../ageAssurance/data')
+jest.mock('../../../ageAssurance/state', () => ({
+  getAndComputeAgeAssuranceState: () => ({}),
+}))
 jest.mock('#/lib/notifications/notifications', () => ({
   unregisterPushToken(_agents: BskyAgent[]) {
     return Promise.resolve()
